tests(api): move test cases that break API contract to its own folder
