+++ tests/configs/public_api.defconfig
@@ -1,0 +1,1 @@
+# CONFIG_LV_CHECK_ARG_ASSERT_ON_FAIL is not set
