+++ tests/src/test_cases/refr/test_refr_api.c
@@ -190,14 +190,4 @@ void test_refr_display_deleted_in_refr_start(void)
     TEST_ASSERT_NULL(refr_ctx.disp);
 }
 
-void test_refr_obj_redraw_invalid_arguments(void)
-{
-    refr_disp_create(64, 64, LV_COLOR_FORMAT_XRGB8888, LV_DISPLAY_RENDER_MODE_DIRECT, 1, 64);
-    lv_layer_t layer;
-    lv_obj_redraw(NULL, refr_screen());
-    lv_obj_redraw(&layer, NULL);
-    lv_obj_refr(NULL, refr_screen());
-    lv_obj_refr(&layer, NULL);
-}
-
 #endif /*LV_BUILD_TEST*/
