+++ tests/src/test_cases/refr/test_refr_top_obj.c
@@ -15,14 +15,6 @@ void tearDown(void)
     refr_disp_delete();
 }
 
-void test_refr_top_obj_invalid_arguments(void)
-{
-    refr_disp_create(64, 64, LV_COLOR_FORMAT_XRGB8888, LV_DISPLAY_RENDER_MODE_DIRECT, 1, 64);
-    lv_area_t area = {0, 0, 9, 9};
-    TEST_ASSERT_NULL(lv_refr_get_top_obj(NULL, refr_screen()));
-    TEST_ASSERT_NULL(lv_refr_get_top_obj(&area, NULL));
-}
-
 void test_refr_top_obj_finds_the_covering_object(void)
 {
     refr_disp_create(64, 64, LV_COLOR_FORMAT_XRGB8888, LV_DISPLAY_RENDER_MODE_DIRECT, 1, 64);
