+++ tests/src/test_cases/test_display.c
@@ -15,7 +15,9 @@ void setUp(void)
 
 void tearDown(void)
 {
+#if LV_DRAW_TRANSFORM_USE_MATRIX
     lv_display_set_matrix_rotation(lv_display_get_default(), false);
+#endif
     lv_obj_clean(lv_screen_active());
     lv_obj_clean(lv_layer_top());
     lv_obj_clean(lv_layer_sys());
@@ -395,13 +397,6 @@ void test_display_resolution(void)
     lv_display_set_offset(lv_display_get_default(), 1, 2);
     test_display_resolution_full_rotation(disp, 32, 48, 24, 32, 1, 2);
 
-    /* Test NULL default display, should not affect the display */
-    lv_display_set_default(NULL);
-    lv_display_set_resolution(lv_display_get_default(), 2, 3);
-    lv_display_set_physical_resolution(lv_display_get_default(), 4, 5);
-    lv_display_set_offset(lv_display_get_default(), 6, 7);
-    test_display_resolution_full_rotation(disp, 32, 48, 24, 32, 1, 2);
-    test_display_resolution_full_rotation(lv_display_get_default(), 0, 0, 0, 0, 0, 0);
 
     /* Restore default display */
     lv_display_set_default(disp_def);
@@ -450,16 +445,6 @@ void test_display_dpi_tile_cnt_antialiasing(void)
     lv_display_set_tile_cnt(lv_display_get_default(), 20);
     TEST_ASSERT_EQUAL_INT32(20, lv_display_get_tile_cnt(lv_display_get_default()));
 
-    /* Test NULL default display, should not affect the display */
-    lv_display_set_default(NULL);
-    lv_display_set_dpi(lv_display_get_default(), 200);
-    TEST_ASSERT_EQUAL_INT32(LV_DPI_DEF, lv_display_get_dpi(lv_display_get_default()));
-    lv_display_set_antialiasing(lv_display_get_default(), false);
-    TEST_ASSERT_FALSE(lv_display_get_antialiasing(lv_display_get_default()));
-    lv_display_set_antialiasing(lv_display_get_default(), true);
-    TEST_ASSERT_FALSE(lv_display_get_antialiasing(lv_display_get_default()));
-    lv_display_set_tile_cnt(lv_display_get_default(), 20);
-    TEST_ASSERT_EQUAL_INT32(0, lv_display_get_tile_cnt(lv_display_get_default()));
 
     /* Restore default display */
     lv_display_set_default(disp_def);
@@ -554,12 +539,6 @@ void test_display_layers(void)
     TEST_ASSERT_NOT_NULL(lv_display_get_layer_sys(lv_display_get_default()));
     TEST_ASSERT_NOT_NULL(lv_display_get_layer_bottom(lv_display_get_default()));
 
-    lv_display_set_default(NULL);
-    TEST_ASSERT_NULL(lv_display_get_screen_active(lv_display_get_default()));
-    TEST_ASSERT_NULL(lv_display_get_screen_prev(lv_display_get_default()));
-    TEST_ASSERT_NULL(lv_display_get_layer_top(lv_display_get_default()));
-    TEST_ASSERT_NULL(lv_display_get_layer_sys(lv_display_get_default()));
-    TEST_ASSERT_NULL(lv_display_get_layer_bottom(lv_display_get_default()));
 
     lv_display_set_default(disp_def);
 }
@@ -576,11 +555,6 @@ void test_display_active_time(void)
     lv_tick_inc(1000);
     TEST_ASSERT_EQUAL_UINT32(1000, lv_display_get_inactive_time(lv_display_get_default()));
 
-    /* Test NULL default display, should not affect the display */
-    lv_display_set_default(NULL);
-    lv_display_trigger_activity(lv_display_get_default());
-    lv_tick_inc(1000);
-    TEST_ASSERT_EQUAL_UINT32(2000, lv_display_get_inactive_time(lv_display_get_default()));
 
     lv_display_set_default(disp_def);
 }
