+++ tests/src/test_cases/test_snapshot.c
@@ -314,22 +314,6 @@ void test_snapshot_reshape_draw_buf_zero_size_object(void)
     lv_draw_buf_destroy(draw_buf);
 }
 
-void test_snapshot_reshape_draw_buf_invalid_params(void)
-{
-    lv_obj_t * obj = lv_obj_create(lv_screen_active());
-    lv_obj_set_size(obj, 50, 40);
-
-    /* Create initial draw buffer */
-    lv_draw_buf_t * draw_buf = lv_draw_buf_create(50, 40, LV_COLOR_FORMAT_ARGB8888, LV_STRIDE_AUTO);
-    TEST_ASSERT_NOT_NULL(draw_buf);
-
-    /* Only test NULL buffer, not NULL object (function asserts obj != NULL internally) */
-    lv_result_t result = lv_snapshot_reshape_draw_buf(obj, NULL);
-    TEST_ASSERT_EQUAL(LV_RESULT_INVALID, result);
-
-    lv_draw_buf_destroy(draw_buf);
-}
-
 void test_snapshot_reshape_draw_buf_reshape_failure(void)
 {
     lv_obj_t * obj = lv_obj_create(lv_screen_active());
