+++ tests/src/test_cases/test_translation.c
@@ -181,25 +181,6 @@ void test_static_pack_rejects_add_language(void)
     TEST_ASSERT_EQUAL(LV_RESULT_INVALID, lv_translation_add_language(pack, "fr"));
 }
 
-void test_static_pack_rejects_add_tag(void)
-{
-    lv_translation_pack_t * pack = add_animals_pack();
-    TEST_ASSERT_NULL(lv_translation_add_tag(pack, "zebra"));
-}
-
-void test_static_pack_rejects_set_tag_translation(void)
-{
-    lv_translation_pack_t * static_pack = add_animals_pack();
-
-    /* A valid tag descriptor is needed, borrow one from a dynamic pack */
-    lv_translation_pack_t * dynamic_pack = lv_translation_add_dynamic();
-    TEST_ASSERT_EQUAL(LV_RESULT_OK, lv_translation_add_language(dynamic_pack, "en"));
-    lv_translation_tag_dsc_t * tag = lv_translation_add_tag(dynamic_pack, "cat");
-    TEST_ASSERT_NOT_NULL(tag);
-
-    TEST_ASSERT_EQUAL(LV_RESULT_INVALID, lv_translation_set_tag_translation(static_pack, tag, 0, "The Cat"));
-}
-
 
 /**********************
  *   DYNAMIC PACKS
@@ -243,17 +224,6 @@ void test_set_tag_translation_overwrites_the_previous_one(void)
 
     TEST_ASSERT_EQUAL(LV_RESULT_OK, lv_translation_set_tag_translation(pack, tag, 0, "The Kitten"));
     TEST_ASSERT_EQUAL_STRING("The Kitten", lv_tr("cat"));
-}
-
-void test_set_tag_translation_rejects_out_of_range_language(void)
-{
-    lv_translation_pack_t * pack = lv_translation_add_dynamic();
-    TEST_ASSERT_EQUAL(LV_RESULT_OK, lv_translation_add_language(pack, "en"));
-
-    lv_translation_tag_dsc_t * tag = lv_translation_add_tag(pack, "cat");
-    TEST_ASSERT_NOT_NULL(tag);
-
-    TEST_ASSERT_EQUAL(LV_RESULT_INVALID, lv_translation_set_tag_translation(pack, tag, 1, "Die Katze"));
 }
 
 void test_deinit_releases_dynamic_packs(void)
