+++ tests/src/test_cases/widgets/test_table.c
@@ -136,9 +136,6 @@ void test_table_should_wrap_long_texts(void)
 
 static void draw_part_event_cb(lv_event_t * e)
 {
-    /* Test lv_event_get_invalidated_area error handling. */
-    TEST_ASSERT_NULL(lv_event_get_invalidated_area(e));
-
     lv_draw_task_t * draw_task = lv_event_get_draw_task(e);
     lv_draw_dsc_base_t * base_dsc = draw_task->draw_dsc;
     /*If the cells are drawn...*/
